With Alien
I changed some identifiers to make it more clear, and kept most of the alien methods, but deleted some to implement it in InvadersGame class. The alien should appear, but does not move yet. I thought that with each drawcurrentstate it would increment its x and y coordinate and the way it switches sides is if the row is even it goes right if it is odd it goes left etc..