separated logic and GUI via a middle controller class
logic is handled by a separate class, GUI class only handles drawings and the controller handles all input/action events. The controller class takes input and passes it to logic which then processes input, updates the objects (ship, aliens, etc) locations. the updated locations are then passed back to the controller which then passes it to the GUI to handle the drawing. the process repeats until game is ended.

I also added a RunMe file which is the execution point of the program. I just like having the main outside for some reason 😅