some in progress thoughts
this stuff doesn't compile I was trying to combine alien shots and player shot together so putting my thoughts out there on it. I also copy pasted all the barrier stuff that was in player shot to barrier class itself. I put some documentation on what I was trying to do

maybe it'll come in useful