basic scoring/file io + clean up
scoring is based on time duration (which kind of doesnt make sense for game over so may have to fix that). it then appends the score to a score log.

also did some cleanup of the code, mostly narrowing down privacy of methods. i may have deleted some stuff that you guys may have needed when i was cleaning up dead code so my bad for that but the project runs fine.